requirements.txt
Django>=4.2,<5
djangorestframework>=3.15
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
django-environ>=0.10
python-dotenv>=1.0.0
django-cors-headers>=4.5.0
gunicorn>=21.2.0
whitenoise>=6.6.0